Create documentations for all python scripts as part of NGS pipeline
Added markdown (md) files to provide thorough and clear documentation for all python scripts employed in the Next-Generation Sequencing (NGS) analysis pipeline. The documentation includes an overview, dependencies, functions, usage, output, and some examples for each script, thus providing a detailed understanding of the functionality and usage of each script in the pipeline.