Add new file 'docs/out.gv' for visualized graph generation
This change introduces a new file 'out.gv' in the 'docs' directory. This file is used for the generation of visualized directed graphs, specifically for tracking the function calls and their dependencies in certain modules of the project. Different shapes and colors represent different types of functions and their interactions.